Move walkers under `assets` package
The walkers are strongly related to assets and there is no reason to
put them in a generic "util" package.
